+++ Extensions/IntuneTools.psm1
@@ -22,7 +22,7 @@ $global:EMToolsViewObject = $null
 
 function Get-ModuleVersion
 {
-    '1.0.3'
+    '1.0.4'
 }
 
 function Invoke-InitializeModule
@@ -82,7 +82,7 @@ function Add-EMToolsViewItem
             Activating = { Invoke-EMToolsActivatingView }
             Authentication = (Get-MSALAuthenticationObject)
             Authenticate = { Invoke-EMToolsAuthenticateToMSAL }
-            AppInfo = (Get-GraphAppInfo "EM" "d1ddf0e4-d672-4dae-b554-9d5bdfd93547")
+            AppInfo = (Get-GraphAppInfo "EMAzureApp" "d1ddf0e4-d672-4dae-b554-9d5bdfd93547")
             SaveSettings = { Invoke-EMSaveSettings }
             Permissions = @()
         })